Revert "Moved files into folders"
This reverts commit 804867f84fd2a8947d9c2050f78f3998801cd018.
