+++ intake.py
